Be

See merge request s11-webmobile1-sub2/S11P12B106!181

+++ BE/pet-bridge/src/main/java/site/petbridge/domain/contract/service/ContractServiceImpl.java
@@ -39,7 +39,7 @@ public class ContractServiceImpl implements ContractService {
 						contract,
 						userRepository.findById(contract.getContractorId()).get(),
 						userRepository.findById(contract.getContracteeId()).get(),
-						animalRepository.findById(contract.getAnimalId()).get()
+						animalRepository.findById((long) contract.getAnimalId()).get()
 					)
 				)
 				.collect(Collectors.toList())
@@ -54,7 +54,7 @@ public class ContractServiceImpl implements ContractService {
 			contract,
 			userRepository.findById(contract.getContractorId()).get(),
 			userRepository.findById(contract.getContracteeId()).get(),
-			animalRepository.findById(contract.getAnimalId()).get(),
+			animalRepository.findById((long) contract.getAnimalId()).get(),
 			contractCheckRepository.findByContractId(id)
 		));
